Edit solution for 0x13-1, handle unsigned long int inputs

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int greatest;
+	unsigned long int greatest;
 
 	if (n == 0)
 	{
